# specs/001-backend-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Beautique Store Backend API
  description: |
    RESTful API for Beautique Store e-commerce platform.

    ## Authentication
    - Public endpoints: No authentication required
    - Admin endpoints (/api/admin/*): JWT Bearer token required

    ## Payment Workflow
    Phase 1 uses manual payment verification via WhatsApp.
    No payment gateway integration.
  version: 1.0.0
  contact:
    name: Beautique Store
    email: admin@beautique.com
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://beautique-api.hf.space
    description: Production server (Hugging Face Spaces)
...
paths:
...
  /api/orders:
    post:
      tags: [Orders]
      summary: Create new order
      description: |
        Creates a new order with customer details and cart items.
        Returns generated Order ID (format: BQ-YYYYMMDD-XXX).
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /api/admin/login:
    post:
      tags: [Admin Auth]
      summary: Admin login
      description: Authenticates admin and returns JWT token
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLogin'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminLoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/admin/orders:
    get:
      tags: [Admin Orders]
      summary: List all orders
      description: Returns paginated list of orders with filters
      operationId: listOrders
      security:
        - BearerAuth: []
      parameters:
        - name: payment_status
          in: query
          schema:
            $ref: '#/components/schemas/PaymentStatus'
        - name: order_status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/admin/orders/{id}:
    get:
      tags: [Admin Orders]
      summary: Get order details
      description: Returns complete order information
      operationId: getOrder
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/orders/{id}/status:
    patch:
      tags: [Admin Orders]
      summary: Update order status
      description: Updates payment, order, or delivery status and tracking notes
      operationId: updateOrderStatus
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderStatusUpdate'
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/orders/export:
    get:
      tags: [Admin Orders]
      summary: Export orders
      description: Returns orders as CSV file
      operationId: exportOrders
      security:
        - BearerAuth: []
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PaymentMethod:
      type: string
      enum:
        - Easypaisa
        - Meezan Bank
        - International Bank
    PaymentStatus:
      type: string
      enum:
        - Pending
        - Paid
        - Verified
    OrderStatus:
      type: string
      enum:
        - Received
        - Processing
        - Ready
        - Delivered
        - Cancelled
    DeliveryStatus:
      type: string
      enum:
        - Not Started
        - In Progress
        - Out for Delivery
        - Delivered

    # ============================================
    # PRODUCT SCHEMAS
    # ============================================
...
    OrderItem:
      type: object
      required:
        - product_id
        - name
        - size
        - color
        - quantity
        - price
      properties:
        product_id:
          type: string
          format: uuid
        name:
          type: string
        size:
          type: string
        color:
          type: string
        quantity:
          type: integer
          minimum: 1
        price:
          type: number
          minimum: 0
    OrderCreate:
      type: object
      required:
        - customer_name
        - phone
        - whatsapp
        - address
        - city
        - items
        - payment_method
      properties:
        customer_name:
          type: string
          minLength: 3
        phone:
          type: string
          pattern: '^[\+]?[(]?[0-9]{1,4}[)]?[-\s\.]?[(]?[0-9]{1,4}[)]?[-\s\.]?[0-9]{1,9}$'
        whatsapp:
          type: string
          pattern: '^[\+]?[(]?[0-9]{1,4}[)]?[-\s\.]?[(]?[0-9]{1,4}[)]?[-\s\.]?[0-9]{1,9}$'
        email:
          type: string
          format: email
        address:
          type: string
          minLength: 10
        city:
          type: string
          minLength: 2
        country:
          type: string
        notes:
          type: string
          maxLength: 500
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
          minItems: 1
        payment_method:
          $ref: '#/components/schemas/PaymentMethod'
    OrderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          example: BQ-20260118-001
        customer_name:
          type: string
        total_amount:
          type: number
        payment_status:
          $ref: '#/components/schemas/PaymentStatus'
        order_status:
          $ref: '#/components/schemas/OrderStatus'
        created_at:
          type: string
          format: date-time
...
    OrderDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
        customer_name:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        total_amount:
          type: number
        payment_method:
          $ref: '#/components/schemas/PaymentMethod'
        payment_status:
          $ref: '#/components/schemas/PaymentStatus'
        order_status:
          $ref: '#/components/schemas/OrderStatus'
        delivery_status:
          $ref: '#/components/schemas/DeliveryStatus'
        estimated_delivery:
          type: string
          format: date
        tracking_notes:
          type: string
        order_date:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderStatusUpdate:
      type: object
      properties:
        payment_status:
          $ref: '#/components/schemas/PaymentStatus'
        order_status:
          $ref: '#/components/schemas/OrderStatus'
        delivery_status:
          $ref: '#/components/schemas/DeliveryStatus'
        estimated_delivery:
          type: string
          format: date
        tracking_notes:
          type: string
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              order_id:
                type: string
              customer_name:
                type: string
              total_amount:
                type: number
              payment_status:
                $ref: '#/components/schemas/PaymentStatus'
              order_status:
                $ref: '#/components/schemas/OrderStatus'
              created_at:
                type: string
                format: date-time
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer

    # ============================================
    # ADMIN SCHEMAS
    # ============================================
    AdminLogin:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    AdminLoginResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          example: 86400
...
    Error:
      type: object
      properties:
        detail:
          type: string
    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string
  responses:
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Invalid or expired token
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Resource not found
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'